HTBHF-1853 move row data transformations to separate file
