+++ src/web/routes/application/check/get.js
@@ -1,7 +1,6 @@
-const { pipe, map, filter, flatten, isNil } = require('ramda')
-const { notIsNil } = require('../../../../common/predicates')
 const { stateMachine, states, actions } = require('../common/state-machine')
 const { getPreviousPath } = require('../common/get-previous-path')
+const { getFlattenedRowData } = require('./get-row-data')
 
 const pageContent = ({ translate }) => ({
   title: translate('check.title'),
@@ -13,23 +12,6 @@ const pageContent = ({ translate }) => ({
   aboutYou: translate('check.aboutYou'),
   aboutYourChildren: translate('check.aboutYourChildren')
 })
-
-const combinePathWithRow = (path) => (row) => ({
-  ...row,
-  path
-})
-
-const getFlattenedRowData = (req) => pipe(map(getRowData(req)), filter(notIsNil), flatten)
-
-const getRowData = (req) => (step) => {
-  if (isNil(step.contentSummary)) {
-    return null
-  }
-  const result = step.contentSummary(req)
-  const applyPathToRow = combinePathWithRow(step.path)
-
-  return Array.isArray(result) ? result.map(applyPathToRow) : applyPathToRow(result)
-}
 
 // a step is navigable if it hasn't defined an isNavigable function.
 const stepIsNavigable = (step, session) => !step.hasOwnProperty('isNavigable') || step.isNavigable(session)
@@ -57,7 +39,5 @@ const getCheck = (steps) => (req, res) => {
 
 module.exports = {
   getCheck,
-  getRowData,
-  getFlattenedRowData,
   getLastNavigablePath
 }
